feat: 앱 상태 모듈로 분리 & Detail 상태 추가
